RENAME FILES TO USE SNAKE_CASE, ADD METHOD FILTER, CHECK LINTERS
